# backend/openapi.yaml
openapi: 3.0.3
info:
  title: TaskFlow API
  description: |
    Comprehensive REST API for TaskFlow task management system.
    
    Features:
    - JWT-based authentication with refresh tokens
    - Complete task CRUD operations with advanced filtering
    - Full-text search across tasks, categories, and tags
    - Real-time notifications and collaboration features
    - Analytics and productivity insights
    - Bulk operations and data import/export
    
    ## Authentication
    All endpoints (except auth endpoints) require a valid JWT Bearer token in the Authorization header.
    
    ## Rate Limiting
    - Default: 100 requests per minute per user
    - Burst: 20 requests per 10 seconds
    - Headers: X-RateLimit-Limit, X-RateLimit-Remaining, X-RateLimit-Reset
    
    ## Error Handling
    All errors follow a consistent format with success/error indicators, error codes, and detailed messages.
  version: 1.0.0
  contact:
    name: TaskFlow API Support
    email: api-support@taskflow.com
    url: https://docs.taskflow.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://taskflow.com/terms
servers:
  - url: https://api.taskflow.com/v1
    description: Production server
  - url: https://staging-api.taskflow.com/v1
    description: Staging server
  - url: http://localhost:3000/v1
    description: Development server
security:
  - bearerAuth: []
paths:
...
  /tasks:
    get:
      tags:
        - Tasks
      summary: Get tasks
      description: Get user's tasks with filtering, search, and pagination
      parameters:
        - name: page
          in: query
          description: Page number
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Items per page
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: status
          in: query
          description: Filter by task status
          schema:
            $ref: '#/components/schemas/TaskStatus'
        - name: priority
          in: query
          description: Filter by task priority
          schema:
            $ref: '#/components/schemas/TaskPriority'
        - name: category_id
          in: query
          description: Filter by category ID
          schema:
            type: string
            format: uuid
        - name: tags
          in: query
          description: Filter by tags (comma-separated)
          schema:
            type: string
            example: "urgent,work"
        - name: due_date_from
          in: query
          description: Tasks due after this date
          schema:
            type: string
            format: date-time
        - name: due_date_to
          in: query
          description: Tasks due before this date
          schema:
            type: string
            format: date-time
        - name: search
          in: query
          description: Full-text search query
          schema:
            type: string
        - name: sort
          in: query
          description: Sort field
          schema:
            type: string
            enum: [created_at, updated_at, due_date, title, priority, status]
            default: created_at
        - name: order
          in: query
          description: Sort order
          schema:
            type: string
            enum: [asc, desc]
            default: desc
        - name: include_completed
          in: query
          description: Include completed tasks
          schema:
            type: boolean
            default: false
        - name: overdue_only
          in: query
          description: Only show overdue tasks
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Tasks retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TasksResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Tasks
      summary: Create task
      description: Create a new task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /tasks/{id}:
    parameters:
      - name: id
        in: path
        required: true
        description: Task ID
        schema:
          type: string
          format: uuid

    get:
      tags:
        - Tasks
      summary: Get task by ID
      description: Get detailed task information including comments and history
      responses:
        '200':
          description: Task retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DetailedTaskResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags:
        - Tasks
      summary: Update task
      description: Update task information
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTaskRequest'
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags:
        - Tasks
      summary: Partially update task
      description: Update specific task fields
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchTaskRequest'
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - Tasks
      summary: Delete task
      description: Soft delete a task
      responses:
        '200':
          description: Task deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ===============================
  # Category Management Endpoints
  # ===============================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token
  schemas:
...
    TaskStatus:
      type: string
      enum: [pending, in_progress, completed, archived]
      description: Task status enumeration
    TaskPriority:
      type: string
      enum: [high, medium, low, none]
      description: Task priority enumeration
...
    BaseResponse:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
          description: Indicates if the request was successful
    SuccessResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            message:
              type: string
              description: Success message
    ErrorResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          required:
            - error
          properties:
            success:
              type: boolean
              enum: [false]
            error:
              type: object
              required:
                - code
                - message
              properties:
                code:
                  type: string
                  description: Error code
                message:
                  type: string
                  description: Error message
                details:
                  type: array
                  items:
                    type: object
                    properties:
                      field:
                        type: string
                      message:
                        type: string
                request_id:
                  type: string
                  description: Unique request identifier
    Pagination:
      type: object
      properties:
        current_page:
          type: integer
          minimum: 1
        per_page:
          type: integer
          minimum: 1
        total:
          type: integer
          minimum: 0
        total_pages:
          type: integer
          minimum: 0
        has_next:
          type: boolean
        has_prev:
          type: boolean

    # ===============================
    # User Schemas
    # ===============================
...
    Task:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Task unique identifier
        title:
          type: string
          maxLength: 255
          description: Task title
        description:
          type: string
          nullable: true
          description: Task description
        status:
          $ref: '#/components/schemas/TaskStatus'
        priority:
          $ref: '#/components/schemas/TaskPriority'
        due_date:
          type: string
          format: date-time
          nullable: true
          description: Task due date
        reminder_date:
          type: string
          format: date-time
          nullable: true
          description: Reminder date
        start_date:
          type: string
          format: date-time
          nullable: true
          description: Task start date
        estimated_minutes:
          type: integer
          minimum: 1
          nullable: true
          description: Estimated time in minutes
        actual_minutes:
          type: integer
          minimum: 0
          nullable: true
          description: Actual time spent in minutes
        completed_at:
          type: string
          format: date-time
          nullable: true
          description: Completion timestamp
        tags:
          type: array
          items:
            type: string
          description: Task tags
        category:
          $ref: '#/components/schemas/Category'
          nullable: true
        metadata:
          type: object
          description: Additional task metadata
        created_at:
          type: string
          format: date-time
          description: Creation timestamp
        updated_at:
          type: string
          format: date-time
          description: Last update timestamp

    # ===============================
    # Category Schemas
    # ===============================
    Category:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Category unique identifier
        name:
          type: string
          maxLength: 100
          description: Category name
        description:
          type: string
          nullable: true
          description: Category description
        color:
          type: string
          pattern: '^#[0-9A-Fa-f]{6}$'
          description: Hex color code
        icon:
          type: string
          description: Icon identifier
        is_default:
          type: boolean
          description: Default category flag
        sort_order:
          type: integer
          description: Display order
        task_count:
          type: integer
          minimum: 0
          description: Number of tasks in category
        completed_task_count:
          type: integer
          minimum: 0
          description: Number of completed tasks
        created_at:
          type: string
          format: date-time
          description: Creation timestamp
        updated_at:
          type: string
          format: date-time
          description: Last update timestamp

    # ===============================
    # Request Schemas
    # ===============================
...
    CreateTaskRequest:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          maxLength: 255
          description: Task title
        description:
          type: string
          description: Task description
        priority:
          $ref: '#/components/schemas/TaskPriority'
        status:
          $ref: '#/components/schemas/TaskStatus'
        category_id:
          type: string
          format: uuid
          description: Category ID
        due_date:
          type: string
          format: date-time
          description: Task due date
        reminder_date:
          type: string
          format: date-time
          description: Reminder date
        start_date:
          type: string
          format: date-time
          description: Task start date
        estimated_minutes:
          type: integer
          minimum: 1
          description: Estimated time in minutes
        tags:
          type: array
          items:
            type: string
          description: Task tags
        metadata:
          type: object
          description: Additional metadata
    UpdateTaskRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 255
        description:
          type: string
        priority:
          $ref: '#/components/schemas/TaskPriority'
        status:
          $ref: '#/components/schemas/TaskStatus'
        category_id:
          type: string
          format: uuid
        due_date:
          type: string
          format: date-time
        reminder_date:
          type: string
          format: date-time
        start_date:
          type: string
          format: date-time
        estimated_minutes:
          type: integer
          minimum: 1
        actual_minutes:
          type: integer
          minimum: 0
        tags:
          type: array
          items:
            type: string
        metadata:
          type: object
    PatchTaskRequest:
      type: object
      properties:
        status:
          $ref: '#/components/schemas/TaskStatus'
        priority:
          $ref: '#/components/schemas/TaskPriority'
        actual_minutes:
          type: integer
          minimum: 0
...
    TaskResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/Task'
    TasksResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                tasks:
                  type: array
                  items:
                    $ref: '#/components/schemas/Task'
                pagination:
                  $ref: '#/components/schemas/Pagination'
                filters_applied:
                  type: object
                  description: Applied filters
                summary:
                  type: object
                  properties:
                    total_tasks:
                      type: integer
                    pending:
                      type: integer
                    in_progress:
                      type: integer
                    completed:
                      type: integer
                    overdue:
                      type: integer
    DetailedTaskResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              allOf:
                - $ref: '#/components/schemas/Task'
                - type: object
                  properties:
                    assignments:
                      type: array
                      items:
                        type: object
                    comments:
                      type: array
                      items:
                        type: object
                    history:
                      type: array
                      items:
                        type: object
...
  responses:
    BadRequest:
      description: Bad request - invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: "VALIDATION_ERROR"
              message: "Invalid input data"
              details:
                - field: "email"
                  message: "Invalid email format"
              request_id: "req-550e8400-e29b-41d4"
    Unauthorized:
      description: Unauthorized - invalid or missing authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: "AUTHENTICATION_ERROR"
              message: "Invalid or expired token"
              request_id: "req-550e8400-e29b-41d4"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: "RESOURCE_NOT_FOUND"
              message: "Task not found"
              request_id: "req-550e8400-e29b-41d4"
...